feat(core): the cast sub-phase in Rust — official pick order, targeting and effects (NML-1073 M1-3b)
Ports BattleSim._cast_phase / _pick_cast / _best_spell_target /
_apply_cast_effect (battle_sim.gd:856-984): the expectation path walks all
three D3 faces at weight 1/3, each face taking the first spell in
AiSpell.official_pick_order (ai_spell.gd:305-312) that is modelled, affordable
and has a legal target; the FIRST face that picks names the attempt and pays
its threshold. Damage lands as the 1/6-scaled expectation, a modifier as a
scaled stamp on the target's mods — with the "beneficiary: attackers" split
that keeps an attacker's modifier out of the bearer's own net. With the seam
ON the shoot branch drops the legacy spell rider, exactly as the GDScript does.

TWO TYPE/INPUT FACTS the port had to fix, both found by the gate, not by
reading:
  * Mods.hit/def/morale are FLOATS. _apply_cast_effect adds `landed * modifier`
    and `landed` is a third of a half, so the M1-2 integer typing would have
    failed to even parse a corpus where a debuff landed.
  * _best_spell_target probes _los_clear with the POST-MOVE centres. The plain
    state carries only state_before's PRE-move los_pairs, so a caster that
    rushed into the open was scored blind: 41 of 93 recorded casts came out
    wrong. Node::cast_los() supplies the recorded post-move row (the same
    recorded-Callable-answer trick cover_dest uses for terrain), which closes
    them all. The caveat is stated in the doc comment, not hidden: the row is
    read off state_after, i.e. the end of the whole activation.

GATE (parity binary, cast corpus ~/selfplay_out/m1_3/run_cast, recorded with
NML_SIM_CAST=1 NML_SIM_SPACING=1): HOLD 550/550, ADVANCE 376/376, RUSH
550/550, CHARGE 524/524 = 2000/2000 exact, 0 mismatched; GATE A 2000/2000
within 1e-9. The three cast-off corpora stay 2000/2000 — 8000 recorded nodes
green in total.

cargo test --release: 21 green (3 unit + 18 parity). New fixture nodes_cast
carries EVERY node of that game whose activation spent a token — 93 of 2000,
31 each on HOLD, RUSH and CHARGE. That split is the point of NML-1069: the
legacy rider only cast inside a shoot pick, so a rushing or charging caster
never cast at all.

Red-green, measured on this code: CAST_BASE_TARGET 4->5, the D3 weight 1/3->1/2
and the official_pick_order start offset each redden the cast gate; dropping
the post-move sight row reddens exactly 41 nodes; running the fixture with the
seam off reddens all 93.

COVERAGE HOLES, named rather than papered over: this corpus never lands a
buff/debuff MODIFIER (mods never moved in 2000 nodes), so the modifier half of
_apply_cast_effect is implemented but ungated; state_to_plain does not record
`cast_events`, so the event log itself is outside the parity contract.

+++ core/nml-core/src/state.rs
@@ -128,14 +128,19 @@ impl Roster {
 
 /// Per-unit modifier snapshot — `mods` is per clone, `mods_base` is the
 /// capture-time reading and is never written (battle_sim.gd:475-479).
+/// All six are FLOATS, not integers: `BattleSim._apply_cast_effect`
+/// (battle_sim.gd:976-982) adds `landed * modifier`, and `landed` is the D3
+/// weight times the cast chance — a third of a half. A snapshot that has never
+/// been cast on carries plain `0`, which is why the M1-2 port could type the
+/// first three as ints and get away with it.
 #[derive(Debug, Clone, Copy, Default, Deserialize)]
 pub struct Mods {
     #[serde(default)]
-    pub hit: i64,
+    pub hit: f64,
     #[serde(default)]
-    pub def: i64,
+    pub def: f64,
     #[serde(default)]
-    pub morale: i64,
+    pub morale: f64,
     #[serde(default)]
     pub range_in: f64,
     #[serde(default)]
